% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNDVI_GOES_MAIN.R
\name{calculateNDVI_GOES_MAIN}
\alias{calculateNDVI_GOES_MAIN}
\title{Main function to calculate all NDVI values for a day for a list of sites}
\usage{
calculateNDVI_GOES_MAIN(day, siteData, year, TZ, dataPath)
}
\arguments{
\item{day}{The desired day of year}

\item{siteData}{matrix of site data where the sites are in individual rows and the columns are siteName, latitude, longitude, timezone (Note: all timezone values must be the same)}

\item{year}{The desired year}

\item{TZ}{The timezone of the sites in siteData (e.g. eastern US sites have a TZ value of 5).}

\item{dataPath}{The directory where all of the GOES data is located}
}
\description{
Main function to calculate all NDVI values for a day for a list of sites
}
