% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calNDVI.R
\name{calNDVI}
\alias{calNDVI}
\title{Calculates NDVI (normalized difference vegetation index)}
\usage{
calNDVI(R2, R3)
}
\arguments{
\item{R2}{The reflectance factor for the red band (channel 2)}

\item{R3}{The reflectance factor for the near infrared band (channel 3)}
}
\description{
Calculates NDVI (normalized difference vegetation index)
}
