% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDiurnalModel.R
\name{createDiurnalModel}
\alias{createDiurnalModel}
\title{Creates the diurnal jags model object}
\usage{
createDiurnalModel(siteName, data)
}
\arguments{
\item{siteName}{The site name used for file naming}

\item{data}{List that includes the data to be fitted (e.g. data$x and data$y) where x is the vector of hours of the day (minutes and seconds included in the decimal) y is the vector of NDVI values}
}
\description{
Creates the diurnal jags model object
}
