% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyFiles.R
\name{createEmptyFiles}
\alias{createEmptyFiles}
\title{Creates empty files for days missing cloud mask data or radiance data}
\usage{
createEmptyFiles(siteData, year, day)
}
\arguments{
\item{siteData}{matrix of site data where the sites are in individual rows and the columns are siteName, latitude, longitude, timezone}

\item{year}{The desired year}

\item{day}{The desired day of year}
}
\description{
Creates empty files for days missing cloud mask data or radiance data
}
