% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNDVI_sub.R
\name{createNDVI_sub}
\alias{createNDVI_sub}
\title{Calculate NDVI for a specific day and time}
\usage{
createNDVI_sub(siteData, day.time, orbitVersion = NEW, dataPath)
}
\arguments{
\item{siteData}{matrix of site data where the sites are in individual rows and the columns are siteName, latitude, longitude, timezone}

\item{day.time}{The year,day,time (e.g. 20172001020 for 10:20 on the 200th day of 2017)}

\item{orbitVersion}{NEW or OLD depending on the position of the satellite (only days before 17 November 2017 are OLD). Default value is NEW}

\item{dataPath}{The directory where all of the GOES data is located}
}
\description{
Calculate NDVI for a specific day and time
}
