% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diurnalExp.R
\name{diurnalExp}
\alias{diurnalExp}
\title{Calculates the y value sequence for the double diurnal curve}
\usage{
diurnalExp(a, c, k, xseq)
}
\arguments{
\item{a}{One value for the a (shape) parameter}

\item{c}{One value for the c (midday NDVI) parameter}

\item{k}{One value for the k (changepoint) parameter}

\item{xseq}{Vector of x values to calculate y values for}
}
\description{
Calculates the y value sequence for the double diurnal curve
}
