% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getABI_Index.R
\name{getABI_Index}
\alias{getABI_Index}
\title{Determines the index in the ABI matrix for the site}
\usage{
getABI_Index(lat, long, orbitVersion = NEW)
}
\arguments{
\item{lat}{Latitude in decimal degrees}

\item{long}{Longitude in decimal degrees}

\item{orbitVersion}{NEW or OLD depending on the position of the satellite (only days before 17 November 2017 are OLD). Default value is NEW}
}
\description{
Determines the index in the ABI matrix for the site
}
