% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataIndex.R
\name{getDataIndex}
\alias{getDataIndex}
\title{Determines the index in the data matrix for the site}
\source{
Based on PUG_L1b-vol3 pg 15
}
\usage{
getDataIndex(vals, ch, orbitVersion = NEW)
}
\arguments{
\item{vals}{The y,x coordinates on the ABI grid (note: the order is y,x not x,y to be consistent with the ABI grid)}

\item{ch}{The channel number (2 indicating the red band, 3 indicating the near infrared band or NDVI, and ACM indicating the cloud mask)}

\item{orbitVersion}{NEW or OLD depending on the position of the satellite (only days before 17 November 2017 are OLD). Default value is NEW.}
}
\value{
Vector of the i,j index for the product matrices
}
\description{
Determines the index in the data matrix for the site
}
