% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMiddayNDVI.R
\name{modelMiddayNDVI}
\alias{modelMiddayNDVI}
\title{Creates the parameter posterior distributions}
\usage{
modelMiddayNDVI(siteData, year, day)
}
\arguments{
\item{siteData}{matrix of site data where the sites are in individual rows and the columns are siteName, latitude, longitude, timezone}

\item{year}{The desired year}

\item{day}{The desired day of year}
}
\description{
Creates the parameter posterior distributions
}
