% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCI.R
\name{plotCI}
\alias{plotCI}
\title{Plots the fitted curve with credible interval for a day}
\usage{
plotCI(siteName, year, day)
}
\arguments{
\item{siteName}{The site name used for file naming}

\item{year}{The desired year}

\item{day}{The desired day of year}
}
\description{
Plots the fitted curve with credible interval for a day
}
