% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMC_Model.R
\name{runMCMC_Model}
\alias{runMCMC_Model}
\title{Runs the MCMC until chains converge and effective sample size > 5000}
\usage{
runMCMC_Model(j.model, variableNames = c("a", "c", "k", "prec"),
  maxIter = 1e+09, baseNum = 80000, iterSize = 40000, maxGBR = 10,
  ID = "")
}
\arguments{
\item{j.model}{The jags model}

\item{variableNames}{Vector of variable names. Default is c("a","c","k","prec").}

\item{maxIter}{The maximum number of iterations to run to wait for both convergence and enough effective samples.}

\item{baseNum}{The base number of iterations to run before checking the Gelman-Brooks-Rubin value}

\item{iterSize}{The number of iterations to run before checking the Gelman-Brooks-Rubin value again}

\item{maxGBR}{The maximum allowable Gelman-Brooks-Rubin value after the base number of iterations have been run. This can be set low (i.e. 3) if you only want to fit days that converge quickly.}

\item{ID}{Identification information to print out if the model does not converge}
}
\description{
Runs the MCMC until chains converge and effective sample size > 5000
}
