% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMiddayCI.R
\name{calculateMiddayCI}
\alias{calculateMiddayCI}
\title{Calculate credible interval on midday NDVI value}
\usage{
calculateMiddayCI(siteName, year, day, savePath)
}
\arguments{
\item{siteName}{The site name used for file naming}

\item{year}{The desired year}

\item{day}{The desired day of year}

\item{savePath}{The folder where the model output is saved}
}
\description{
Calculate credible interval on midday NDVI value
}
